Check more features and targets in CI when linting
