+++ esp-metadata/Cargo.toml
@@ -8,6 +8,8 @@ repository   = "https://github.com/esp-rs/esp-hal"
 license      = "MIT OR Apache-2.0"
 
 [dependencies]
+anyhow     = "1.0.86"
+clap       = { version = "4.5.4",  features = ["derive"] } # TODO remove this dep here
 basic-toml  = "0.1.9"
 lazy_static = "1.5.0"
 serde       = { version = "1.0.204", features = ["derive"] }
