+++ xtask/src/lib.rs
@@ -9,6 +9,7 @@ use std::{
 use anyhow::{bail, Result};
 use cargo::CargoAction;
 use clap::ValueEnum;
+use esp_metadata::Chip;
 use strum::{Display, EnumIter, IntoEnumIterator as _};
 
 use self::cargo::CargoArgsBuilder;
@@ -50,69 +51,6 @@ pub enum Package {
     HilTest,
     XtensaLx,
     XtensaLxRt,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Display, EnumIter, ValueEnum, serde::Serialize)]
-#[serde(rename_all = "kebab-case")]
-#[strum(serialize_all = "kebab-case")]
-pub enum Chip {
-    Esp32,
-    Esp32c2,
-    Esp32c3,
-    Esp32c6,
-    Esp32h2,
-    Esp32s2,
-    Esp32s3,
-}
-
-impl Chip {
-    pub fn target(&self) -> &str {
-        use Chip::*;
-
-        match self {
-            Esp32 => "xtensa-esp32-none-elf",
-            Esp32c2 | Esp32c3 => "riscv32imc-unknown-none-elf",
-            Esp32c6 | Esp32h2 => "riscv32imac-unknown-none-elf",
-            Esp32s2 => "xtensa-esp32s2-none-elf",
-            Esp32s3 => "xtensa-esp32s3-none-elf",
-        }
-    }
-
-    pub fn has_lp_core(&self) -> bool {
-        use Chip::*;
-
-        matches!(self, Esp32c6 | Esp32s2 | Esp32s3)
-    }
-
-    pub fn lp_target(&self) -> Result<&str> {
-        use Chip::*;
-
-        match self {
-            Esp32c6 => Ok("riscv32imac-unknown-none-elf"),
-            Esp32s2 | Esp32s3 => Ok("riscv32imc-unknown-none-elf"),
-            _ => bail!("Chip does not contain an LP core: '{}'", self),
-        }
-    }
-
-    pub fn pretty_name(&self) -> &str {
-        match self {
-            Chip::Esp32 => "ESP32",
-            Chip::Esp32c2 => "ESP32-C2",
-            Chip::Esp32c3 => "ESP32-C3",
-            Chip::Esp32c6 => "ESP32-C6",
-            Chip::Esp32h2 => "ESP32-H2",
-            Chip::Esp32s2 => "ESP32-S2",
-            Chip::Esp32s3 => "ESP32-S3",
-        }
-    }
-
-    pub fn is_xtensa(&self) -> bool {
-        matches!(self, Chip::Esp32 | Chip::Esp32s2 | Chip::Esp32s3)
-    }
-
-    pub fn is_riscv(&self) -> bool {
-        !self.is_xtensa()
-    }
 }
 
 #[derive(Debug, Default, Clone)]
